Rename `test_data.py`
With the `test_` prefix, Pytest thinks that it's a file containing
tests, which then leads to a warning about no tests being found.
