+++ mavis/test/models/import_records.py
@@ -1,9 +1,9 @@
 from typing import Final, Optional
 
+from ..data import TestData
 from ..generic_constants import actions, escape_characters, properties, wait_time
-from ..playwright_ops import PlaywrightOperations
 from ..mavis_constants import mavis_file_types, Location
-from ..test_data import TestData
+from ..playwright_ops import PlaywrightOperations
 from ..wrappers import get_link_formatted_date_time
 
 from .children import ChildrenPage
