+++ mavis/test/models/programmes.py
@@ -2,6 +2,7 @@ from typing import Final
 
 import pandas as pd
 
+from ..data import TestData
 from ..generic_constants import actions, properties, wait_time
 from ..mavis_constants import (
     Location,
@@ -10,7 +11,6 @@ from ..mavis_constants import (
     Programme,
 )
 from ..playwright_ops import PlaywrightOperations
-from ..test_data import TestData
 from ..wrappers import get_current_datetime, get_link_formatted_date_time
 
 from .children import ChildrenPage
